Fixed a bug in division math where ints were not being promoted to doubles.  This zeroed out some scores that shouldn't have been.
Added a 'last ten day' activity score feature which displays bar graph images corresponding to how active the user has been the last ten days with ten small images.

These are all features I will implement for now; will focus on pretty-ing up the website first.  If more time arises after this, we'll work on other features.

+++ src/java/com/myurlname/stepup/Achievement.java
@@ -24,51 +24,51 @@ public class Achievement implements Serializable {
     private double activityValue; //how much towards the weekly goal is this?
     private static final double CAP = 1.252;
     private static final double MAX_DAILY_VALUE = 0.3334;
-    
+
     public Achievement (Activity activity, int minutes, Intensity intensity,
-                        double score, String notes, Date activityDate, 
+                        double score, String notes, Date activityDate,
                                                           Date recordedDate) {
-        
+
         if (notes != null) {
             notes = StringEscapeUtils.escapeHtml4(notes);
-            this.notes = notes.replace("'", "&#39;");    
+            this.notes = notes.replace("'", "&#39;");
         }
         this.activityDate = activityDate;
         this.recordedDate = recordedDate;
         this.minutes = minutes;
         this.activity = activity;
-        this.intensity = intensity;     
+        this.intensity = intensity;
         this.score = score;
-    }          
-    
+    }
+
     public Achievement (AchievementBean bean) {
-        try { 
+        try {
             activity = new Activity (bean.getActivity());
-            intensity = new Intensity (bean.getIntensity());            
+            intensity = new Intensity (bean.getIntensity());
             minutes = Integer.parseInt(bean.getMinutes());
             score = calculateScore (activity,intensity,minutes);
             if (bean.getNotes() != null) {
                 notes = StringEscapeUtils.escapeHtml4(bean.getNotes());
-                notes = notes.replace("'", "&#39;");    
-            }            
-                        
+                notes = notes.replace("'", "&#39;");
+            }
+
             SimpleDateFormat sdf = new SimpleDateFormat("M/dd/yyyy");
             String strDate = bean.getDateActivity();
             if (strDate != null) {
                 strDate = strDate.replace('-', '/');
                 strDate = strDate.replace('.', '/');
                 strDate = strDate.replace('\\', '/');
-                activityDate = sdf.parse(strDate);                                                            
+                activityDate = sdf.parse(strDate);
             }
             strDate = bean.getDateRecorded();
             if (strDate != null) {
                 strDate = strDate.replace('-', '/');
                 strDate = strDate.replace('.', '/');
                 strDate = strDate.replace('\\', '/');
-                recordedDate = sdf.parse(strDate);                 
-            }    
-            else 
-                recordedDate = new Date();  //not really used, it's set by dbase            
+                recordedDate = sdf.parse(strDate);
+            }
+            else
+                recordedDate = new Date();  //not really used, it's set by dbase
         }
         catch (Exception e) {
             //something went wrong with the inputs, so let's null out stuff
@@ -82,20 +82,20 @@ public class Achievement implements Serializable {
             recordedDate = null;
         }
     }
-    
-    public Achievement () {     
+
+    public Achievement () {
     }
 
     public boolean validate () {
         Date today = new Date();
-        if (minutes <= 0) 
+        if (minutes <= 0)
             return false;
-        if (notes.length() > 200) 
+        if (notes.length() > 200)
             return false;
         if (recordedDate == null)
             return false;
         if (recordedDate.compareTo(today)> 0)
-            return false;        
+            return false;
         if (activityDate == null)
             return false;
         if (activityDate.compareTo(today)> 0)
@@ -103,39 +103,39 @@ public class Achievement implements Serializable {
         if (activity == null)
             return false;
         if (intensity == null)
-            return false;              
-        return true; 
+            return false;
+        return true;
     }
-    
-    private double calculateScore (Activity activity, Intensity intensity, 
+
+    private double calculateScore (Activity activity, Intensity intensity,
                                                                 int minutes) {
         /*the ALGORITHM for calculating the worth of an exercise
         (muscleFactor 1 <=> 10).  We want maximum results when MF is 10.
         This means we take (MuscleFactor / 10)
-        We also have the concept of duration.  Duration required to get full 
+        We also have the concept of duration.  Duration required to get full
         credit is based on intensity.
-        
+
         (minutes/IntensityMinutes)
         Multiply the two to get a score between 0 and 1
         score = muscleFactor/10  *  minutes/intensityMinutes
         Cap scores more than 1 to CAP (this allows light/moderate exercises to
         skip a day (do 4 instead of 5) if they exercise a lot)
-        Examples:  
+        Examples:
         Running at moderate intensity for 20 min = 10/10 * 20/30 = 0.66
         Climbing strenuously for 15 minutes = 8/10 * 15/20 = 0.6
         Hiking lightly for 6 hours = 4/10 * 240/30 = CAP (see top of file)
         */
         int muscleF = activity.getMuscleFactor();
         int minMinutes = intensity.getMinimumMinutes();
-        double tempScore = muscleF / Activity.ACT_MAXMF;
-        tempScore = tempScore * (minutes/intensity.getMinimumMinutes());
+        double tempScore = muscleF / (double) Activity.ACT_MAXMF;
+        tempScore = tempScore * (minutes/ (double) intensity.getMinimumMinutes());
         if (tempScore > CAP)
             tempScore = CAP;
         else if (tempScore < 0)
             tempScore = 0.0;
-        return tempScore;                        
+        return tempScore;
     }
-    
+
     public Activity getActivity() {
         return activity;
     }
@@ -143,10 +143,10 @@ public class Achievement implements Serializable {
     public void setActivity(Activity activity) {
         this.activity = activity;
     }
-    
+
     public void setActivity(String activity) {
         this.activity = new Activity (activity);
-    }    
+    }
 
     public int getMinutes() {
         return minutes;
@@ -163,11 +163,11 @@ public class Achievement implements Serializable {
     public void setIntensity(Intensity intensity) {
         this.intensity = intensity;
     }
-    
+
     public void setIntensity(String intensity) {
         this.intensity = new Intensity(intensity);
-    }    
-    
+    }
+
 
     public String getNotes() {
         return notes;
@@ -176,7 +176,7 @@ public class Achievement implements Serializable {
     public void setNotes(String notes) {
         if (notes != null) {
             notes = StringEscapeUtils.escapeHtml4(notes);
-            this.notes = notes.replace("'", "&#39;");    
+            this.notes = notes.replace("'", "&#39;");
         }
     }
 
@@ -211,20 +211,20 @@ public class Achievement implements Serializable {
     public void setAchievementId(int achievementId) {
         this.achievementId = achievementId;
     }
-    
+
     public String getPrettyPrintActivityDate () {
         SimpleDateFormat sdf = new SimpleDateFormat("MM/dd/yyyy");
-        return sdf.format(this.activityDate);           
-        
+        return sdf.format(this.activityDate);
+
     }
-    
+
     @Override
     public String toString () {
-        String text = 
+        String text =
             String.format("%s: %s for %d minutes at %s intensity",
                            getPrettyPrintActivityDate(),this.activity,
                            this.minutes,this.intensity);
-        
+
         if (!(this.notes == null) && (this.notes.length()>0))
             text = text + " [" + this.notes + "]";
         return text;
@@ -251,7 +251,7 @@ public class Achievement implements Serializable {
     public void setActivityValue(double activityValue) {
         if (activityValue < 0)
                 this.activityValue = 0.0;
-        else if (activityValue >  MAX_DAILY_VALUE) 
+        else if (activityValue >  MAX_DAILY_VALUE)
             this.activityValue = MAX_DAILY_VALUE;
         else
             this.activityValue = activityValue;
